requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bitarray==2.6.0
black==22.6.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
cssselect==1.1.0
cytoolz==0.12.0
demjson3==3.0.5
Django==4.0.7
djangorestframework==3.13.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
flake8==5.0.4
frozenlist==1.3.1
gunicorn==20.1.0
h11==0.13.0
hexbytes==0.3.0
idna==3.3
importlib-metadata==4.12.0
ipfshttpclient==0.8.0a2
joblib==1.2.0
jsonschema==4.16.0
kombu==5.2.4
lru-dict==1.1.8
lxml==4.9.1
Markdown==3.3.7
mccabe==0.7.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.24.1
outcome==1.1.0
pandas==1.5.3
parsimonious==0.8.1
pathspec==0.9.0
platformdirs==2.5.2
postgres==4.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psycopg2-binary==2.9.5
psycopg2-pool==1.1
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
pyflakes==2.5.0
pyOpenSSL==22.0.0
pyquery==1.4.3
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.1
rabbitmq==0.2.0
requests==2.27.1
rlp==2.0.1
scikit-learn==1.2.1
scipy==1.10.0
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.12.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.9
varint==1.0.2
vine==5.0.0
wcwidth==0.2.6
web3==5.30.0
websockets==9.1
wsproto==1.1.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.1
zipp==3.8.0
